requirements.txt
aioredis==1.0.0
asgi-redis==1.0.0
asgiref==2.3.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
channels==2.0.0
channels-api==0.4.0
channels-redis==2.1.1
constantly==15.1.0
daphne==2.1.1
Django==2.0
djangorestframework==3.8.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.6
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2018.4
redis==2.10.6
selenium==3.11.0
six==1.11.0
Twisted==17.9.0
txaio==2.10.0
zope.interface==4.5.0
